Arch for s2sServiceTokenGeneration.
