+++ api/lib/s2sTokenGeneration.ts
@@ -1,0 +1,8 @@
+/**
+ * Return the generated S2s token
+ *
+ * @returns {string}
+ */
+export function generateS2sToken() {
+  return 'token'
+}
