+++ api/openRoutes.ts
@@ -1,5 +1,7 @@
 import * as express from 'express'
-import stateRouter from './states'
+
+// TODO: rename from prdRouter
+import prdRouter from './prd'
 
 // TODO: Not sure if this is needed
 // process.env.NODE_TLS_REJECT_UNAUTHORIZED = '0'
@@ -15,6 +17,6 @@ const router = express.Router({ mergeParams: true })
  *
  * @see local.ts / server.ts
  */
-router.use('/decisions', stateRouter)
+router.use('/decisions', prdRouter)
 
 export default router
